src/main/java/com/dg/s1/bankbook/BankbookDAO.java
package com.dg.s1.bankbook;

import java.util.List;

import org.apache.ibatis.session.SqlSession;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Repository;


@Repository
public class BankbookDAO {
	
	@Autowired
	private SqlSession sqlSession;
	private final String NAMESPACE="com.dg.s1.bankbook.BankbookDAO.";
	
	public int setInsert(BankbookDTO bankbookDTO) {
		return sqlSession.insert(NAMESPACE+"setInsert", bankbookDTO);
	}
	
	public List<BankbookDTO> getList() {
		return sqlSession.selectList(NAMESPACE+"getList");
	}

	public BankbookDTO getSelect(BankbookDTO bankbookDTO) {
		
		return sqlSession.selectOne(NAMESPACE+"getSelect", bankbookDTO);

	}

}
